refactor: move validations to utils and refactor with an abstract class implementation
